Remove auth helper and move onAutStateChaged login to authentication action
