+++ src/js/actions/authentication.js
@@ -1,6 +1,7 @@
-// import { auth, database, authUI } from '../../firebase';
-// import { addUser } from './users';
-import { auth } from '../../firebase';
+import moment from 'moment';
+import { auth, database } from '../../firebase';
+
+const appUsersRef = database.ref('appUsers');
 
 const loggedIn = user => ({
   type: 'LOGIN',
@@ -17,7 +18,7 @@ const loggedOut = () => ({
 
 export const login = () => dispatch => {
   dispatch({ type: 'ATTEMPTING_LOGIN' });
-  // auth.signInWithPopup(googleAuthProvider);
+  // TODO ADD SIGNIN METHOD FROM LOGIN.JSX HERE
 };
 
 export const logout = () => dispatch => {
@@ -29,6 +30,21 @@ export const startListeningToAuthChanges = () => dispatch => {
   auth.onAuthStateChanged(user => {
     if (user) {
       dispatch(loggedIn(user));
+      const appUserRef = appUsersRef.child(user.uid);
+      appUserRef.once('value').then(snapshot => {
+        if (snapshot.val()) return;
+        const date = moment().toDate();
+        const userData = {
+          createdOn: date,
+          displayName: user.displayName,
+          email: user.email,
+          photoURL:
+            user.photoURL ||
+            `https://firebasestorage.googleapis.com/v0/b/dowhop-me.appspot.com/o/assets%2Ficons%2Fprofile-placeholder.png?alt=media&token=04033fce-4c9b-4976-8407-da41981f8046`,
+          uid: user.uid
+        };
+        appUserRef.update(userData);
+      });
     } else {
       dispatch(loggedOut());
     }
