+++ src/js/components/App.jsx
@@ -14,7 +14,6 @@ import Login from './Login';
 import Me from './Me';
 import Profile from './Profile';
 import FourOhFour from './FourOhFour';
-
 import { startListeningToAuthChanges } from '../actions/authentication';
 
 store.dispatch(startListeningToAuthChanges());
