+++ src/js/components/PrivateRoute.jsx
@@ -5,23 +5,19 @@ import { connect } from 'react-redux';
 import { Route, Redirect } from 'react-router-dom';
 import renderMergedProps from '../helpers/renderMergedProps';
 
-const PrivateRoute = ({ component, redirectTo, authentication, ...rest }: Object) => {
-  console.log('PRIVATE ROUTE ->', authentication.status);
-  return (
-    <Route
-      {...rest}
-      render={routeProps =>
-        authentication.isAuthenticated
-          ? renderMergedProps(component, routeProps, rest)
-          : <Redirect
-              to={{
-                pathname: redirectTo,
-                state: { from: routeProps.location }
-              }}
-            />}
-    />
-  );
-};
+const PrivateRoute = ({ component, redirectTo, authentication, ...rest }: Object) =>
+  <Route
+    {...rest}
+    render={routeProps =>
+      authentication.isAuthenticated
+        ? renderMergedProps(component, routeProps, rest)
+        : <Redirect
+            to={{
+              pathname: redirectTo,
+              state: { from: routeProps.location }
+            }}
+          />}
+  />;
 
 const mapStateToProps = ({ authentication }) => ({ authentication });
